% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_dkno.R
\name{cb_dkno}
\alias{cb_dkno}
\title{Detect "Don't Know" and "None" Labels}
\usage{
cb_dkno(data = upanel.meta, num.var, .str_dk = str_dk, .str_no = str_no)
}
\arguments{
\item{data}{a \code{data.frame} object}

\item{num.var}{numvar placeholder}

\item{.str_dk}{str_dk placeholder}

\item{.str_no}{str_no placeholder}
}
\value{
cb_dkno object
}
\description{
This is mainly a helper function for the creation of the codebook table.
It detects whether a given variable has "Don't Know" or "None"  response labels.
}
