% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_mistable.R
\name{cb_mistable}
\alias{cb_mistable}
\title{Codebook Table for missing labels}
\usage{
cb_mistable(metadata, num.var = 1, .miscodes = miscodes)
}
\arguments{
\item{metadata}{a \code{data.frame} object with survey metadata}

\item{num.var}{single numeric value indexing column of variable in metadata (defaults to 1 as missing codes are usually available for the first variable)}

\item{.miscodes}{numeric vector indexing rows of missing labels in metadata (defaults to miscodes)}
}
\value{
table with missing labels
}
\description{
Codebook Table for missing labels
}
