% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_page.R
\name{cb_page}
\alias{cb_page}
\title{Page with Codebook Table for one variabe}
\usage{
cb_page(
  metadata,
  num.var,
  comment = "",
  lbl.space = "1em",
  lblen.space = "1em",
  mis.space = "1em"
)
}
\arguments{
\item{metadata}{a \code{data.frame} object with survey metadata}

\item{num.var}{single numeric value indexing column of variable in metadata}

\item{comment}{optional character argument to add text below table}

\item{lbl.space}{optional argument to adjust table space between variable information and German value labels (defaults to 1em)}

\item{lblen.space}{optional argument to adjust table space between German value labels and English value labels (defaults to 1em)}

\item{mis.space}{optional argument to adjust table space between English value labels and missing labels (defaults to 1em)}
}
\value{
codebook page for one variable in Latex format
}
\description{
Page with Codebook Table for one variabe
}
