% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_pages.R
\name{cb_pages}
\alias{cb_pages}
\title{Pages with Codebook Tables for multiple variables}
\usage{
cb_pages(
  metadata,
  multi.vars,
  comment = "",
  lbl.space = "1em",
  lblen.space = "1em",
  mis.space = "1em"
)
}
\arguments{
\item{metadata}{a \code{data.frame} object with survey metadata}

\item{multi.vars}{numeric vector indexing columns of variables in metadata}

\item{comment}{optional character argument to add text below table}

\item{lbl.space}{optional argument to adjust table space between variable information and German value labels (defaults to 1em)}

\item{lblen.space}{optional argument to adjust table space between German value labels and English value labels (defaults to 1em)}

\item{mis.space}{optional argument to adjust table space between English value labels and missing labels (defaults to 1em)}
}
\value{
codebook pages for multiple variables in Latex format
}
\description{
Pages with Codebook Tables for multiple variables
}
