% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_sumplot.R
\name{cb_sumplot}
\alias{cb_sumplot}
\title{Summary Plot of responses for codebook}
\usage{
cb_sumplot(metadata, response, num.var, na_sep = TRUE, stats)
}
\arguments{
\item{metadata}{a \code{data.frame} object with survey metadata}

\item{response}{a \code{data.frame} object with survey response data}

\item{num.var}{single numeric value indexing column of variable in metadata}

\item{na_sep}{a boolean indicating use of SEP coding rules (defaults to TRUE)}

\item{stats}{a character indicating plotted statistic (count, density)}
}
\value{
plot with counts or density of responses
}
\description{
Summary Plot of responses for codebook
}
