% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_tab.R
\name{cb_tab}
\alias{cb_tab}
\title{Creates a Codebook Table for a given variable (row)}
\usage{
cb_tab(
  data = upanel.meta,
  num.var,
  .meta = meta,
  .codes = codes,
  .codes.en = codes.en,
  .miscodes = miscodes,
  .str_dk = str_dk,
  .str_no = str_no,
  lbl.space = "1em",
  lblen.space = "1em",
  mis.space = "1em"
)
}
\arguments{
\item{data}{a \code{data.frame} object}

\item{num.var}{single numeric value referring to column index}

\item{.meta}{numeric vector referring to row indices of variable information}

\item{.codes}{numeric vector referring to row indices of variable codes and labels}

\item{.codes.en}{numeric vector referring to row indices of variable codes and labels in English}

\item{.miscodes}{miscodes placeholder}

\item{.str_dk}{a character vector including all response options that are coded as -8 ("Don't Know" category)}

\item{.str_no}{a character vector including all response options that are coded as -9 ("None" category)}

\item{lbl.space}{optional argument to adjust table space between variable information and value labels}

\item{lblen.space}{optional argument to adjust table space between value labels and value labels in English}

\item{mis.space}{optional argument to adjust table space between value labels in English and missing labels}
}
\value{
codebook table in Latex format
}
\description{
Creates a Codebook Table for a given variable (row)
}
