% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_tab_multi.R
\name{cb_tab_multi}
\alias{cb_tab_multi}
\title{Creates Page with Table in Codebook for multiple variabes}
\usage{
cb_tab_multi(
  multi.vars,
  data = upanel.meta,
  comment = "",
  lbl.space = "1em",
  lblen.space = "1em",
  mis.space = "1em"
)
}
\arguments{
\item{multi.vars}{numeric vector referring to column indeces}

\item{data}{a \code{data.frame} object}

\item{comment}{optional string argument printed below the codebook table}

\item{lbl.space}{optional argument to adjust table space between variable information and value labels}

\item{lblen.space}{optional argument to adjust table space between value labels and value labels in English}

\item{mis.space}{optional argument to adjust table space between value labels in English and missing labels}
}
\value{
codebook pages for multiple variables in Latex format
}
\description{
Creates Page with Table in Codebook for multiple variabes
}
