% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_tab_single.R
\name{cb_tab_single}
\alias{cb_tab_single}
\title{Creates Page with Table in Codebook for a single variabe}
\usage{
cb_tab_single(
  num.var,
  data = upanel.meta,
  comment = "",
  lbl.space = "1em",
  lblen.space = "1em",
  mis.space = "1em"
)
}
\arguments{
\item{num.var}{num.var placeholder}

\item{data}{a \code{data.frame} object}

\item{comment}{comment placeholder}

\item{lbl.space}{lbl.space placeholder}

\item{lblen.space}{lblen.space placeholder}

\item{mis.space}{mis.space placeholder}
}
\value{
Latex document with Table from Kable<
}
\description{
Creates Page with Table in Codebook for a single variabe
}
