% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstable.R
\name{crosstable}
\alias{crosstable}
\title{Create simple crosstables of n variables}
\usage{
crosstable(vars, df)
}
\arguments{
\item{vars}{vector of variables}

\item{df}{dataframe}
}
\value{
a crosstable
}
\description{
Unfortunately, for the time being this function only works with two numeric variables with values above 0. Updates will follow.
}
\examples{
crosstable(vars = c("gear", "carb"), df = mtcars)
}
