% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_crosstable.R
\name{eda_crosstable}
\alias{eda_crosstable}
\title{Calculates a frequency table for two survey items}
\usage{
eda_crosstable(data = currentwave, items, rm.dk = TRUE, itemnames = NULL)
}
\arguments{
\item{data}{A \code{data.frame} object}

\item{items}{survey items (in form of a vector)}

\item{rm.dk}{a logical value indicating whether "Don't know" values (-8) should be stripped before the computation proceeds}

\item{itemnames}{optional argument for specifying the survey items' names}
}
\value{
a table with relative frequencies
}
\description{
Calculates a frequency table for two survey items
}
\examples{
\dontrun{
eda_crosstable(data = infert, items = c("education", "parity"), rm.dk = FALSE)
}
}
