% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_freqtable.R
\name{eda_freqtable}
\alias{eda_freqtable}
\title{Calculates a frequency table for a single survey item}
\usage{
eda_freqtable(data, item, rm.dk = FALSE)
}
\arguments{
\item{data}{a \code{data.frame} object}

\item{item}{a survey item}

\item{rm.dk}{a logical value indicating whether "Don't know" values (-8) should be stripped before the computation proceeds}
}
\value{
a frequency table (in percent)
}
\description{
Calculates a frequency table for a single survey item
}
\examples{
\dontrun{
eda_freqtable(data = w5, item = w5_q11, rm.dk = FALSE)
}
}
