% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_loadtoenv.R
\name{eda_loadtoenv}
\alias{eda_loadtoenv}
\title{Loading the relevant survey into your environment}
\usage{
eda_loadtoenv(data)
}
\arguments{
\item{data}{A data.frame object}
}
\value{
Loads the relevant object into the global environment
}
\description{
The functions of the package are written in such a way that once the survey is loaded into the global environment, the survey does not need to be specified each time a function is run. Consequently, the first step in the workflow is to use this function to read in the relevant survey (in data.frame format).
}
\examples{
eda_loadtoenv(mtcars)
}
