% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_particpants.R
\name{eda_participants}
\alias{eda_participants}
\title{Helper function to quickly calculated the number of participants for the corresponding survey item}
\usage{
eda_participants(data = currentwave, item, rm.dk = FALSE)
}
\arguments{
\item{data}{a data.frame object}

\item{item}{survey item}

\item{rm.dk}{a logical value indicating whether "Don't know" values (-8) should be stripped before the computation proceeds}
}
\value{
number of particpants
}
\description{
Helper function to quickly calculated the number of participants for the corresponding survey item
}
\examples{
\dontrun{
eda_participants(data = w5, item = w5_q11)
}
}
