% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_percentage.R
\name{helper_percentage}
\alias{helper_percentage}
\title{Helper function to easily calculate/display relative frequencies}
\usage{
helper_percentage(item, digits = 1)
}
\arguments{
\item{item}{a survey item}

\item{digits}{number of digits to be displayed}
}
\value{
relative frequencies of a survey item
}
\description{
Helper function to easily calculate/display relative frequencies
}
