% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_par.R
\name{n_par}
\alias{n_par}
\title{Number of participants for survey item}
\usage{
n_par(data, item, lang = "DE")
}
\arguments{
\item{data}{a \code{data.frame} object}

\item{item}{relevant survey item (i.e. w5_q17)}

\item{lang}{language (German = "DE" (default), English = "EN", anything else goes neutral)}
}
\value{
Number of participants ("Don't knows"-answers included) for item
}
\description{
Number of participants for survey item
To be used with @backref plot_bar_h, @backref plot_bar_v, @backref plot_groupbar_h and @backref plot_groupbar_v.
For @backref plot_multiple_h and @backref plot_rank_v, use the function @backref n_par_by.
}
