% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar_h.R
\name{plot_bar_h}
\alias{plot_bar_h}
\title{Simple horizontal barplot}
\usage{
plot_bar_h(
  data,
  item,
  weights,
  lang = "DE",
  barcolor = "#1F407A",
  barwidth = 0.8,
  texthjust = 2,
  textsize = 8,
  min_textsize = 5
)
}
\arguments{
\item{data}{a \code{data.frame} object}

\item{item}{a survey item}

\item{weights}{optional argument to weight output by survey weights}

\item{lang}{optional argument for language (German = "DE" (default), English = "EN")}

\item{barcolor}{optional argument to set the color of the bar}

\item{barwidth}{optional argument to define the width of the bar}

\item{texthjust}{optional argument to adjust the text's horizontal alignment}

\item{textsize}{optional argument to adjust the text's size}

\item{min_textsize}{optional argument to set the minimum text size}
}
\value{
a horizontal barplot
}
\description{
Simple horizontal barplot
}
