% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_groupbar_h.R
\name{plot_groupbar_h}
\alias{plot_groupbar_h}
\title{Grouped horizontal barplot}
\usage{
plot_groupbar_h(data, item, by, barpadding = 0.1, legendtitle = "", ...)
}
\arguments{
\item{data}{a \code{data.frame} object}

\item{item}{a survey item}

\item{by}{grouping variable}

\item{barpadding}{optional argument to adjust padding between bars}

\item{legendtitle}{optional argument to define a legend title}

\item{...}{further arguments of \code{scale_fill_discrete} like \code{labels}}
}
\value{
a grouped horizontal barplot
}
\description{
Grouped horizontal barplot
}
