% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multiple_h.R
\name{plot_multiple_h}
\alias{plot_multiple_h}
\title{Horizontal barplot for multiple items}
\usage{
plot_multiple_h(
  data,
  item,
  by,
  treat,
  weights,
  item_labels,
  question,
  lang = "DE",
  barwidth = 0.6,
  ncol.wrap = 1,
  textsize = 8,
  min_textsize = 5,
  percent_position = "center",
  legend.pos = "bottom",
  ...
)
}
\arguments{
\item{data}{a \code{data.frame} object}

\item{item}{a \code{vector} listing the survey items}

\item{by}{optional argument for distinction by subgroup}

\item{treat}{optional argument for distinction by treatment group}

\item{weights}{optional argument to weight output by survey weights}

\item{item_labels}{a \code{vector} listing the item labels for the y-axis}

\item{question}{question text}

\item{lang}{optional argument for language (German = "DE" (default), English = "EN")}

\item{barwidth}{optional argument to define the width of the bar}

\item{ncol.wrap}{optional argument to set the number of facet_wrap columns when distinguishing by subgroup or treatment group}

\item{textsize}{optional argument to adjust the text's size}

\item{min_textsize}{optional argument to set the minimum text size}

\item{percent_position}{optional argument to set position of percentage within bar ("center", "left", "top", etc.)}

\item{legend.pos}{optional argument to set position of legend}

\item{...}{further arguments of \code{scale_fill_manual} like \code{labels}}
}
\value{
a vertical barplot for ranking
}
\description{
Horizontal barplot for multiple items
}
