% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rank_v.R
\name{plot_rank_v}
\alias{plot_rank_v}
\title{Vertical barplot to rank items}
\usage{
plot_rank_v(
  data,
  item,
  by = NULL,
  treat = NULL,
  lang = "DE",
  n.items = length(item_labels),
  ncol.wrap = 1,
  textsize = 8,
  min_textsize = 5,
  percent_position = "center",
  barwidth = 0.6,
  legend.pos = "right"
)
}
\arguments{
\item{data}{a \code{data.frame} object}

\item{item}{a vector lisiting the survey items}

\item{by}{optional argument for distinction by subgroup}

\item{treat}{optional argument for distinction by treatment group}

\item{lang}{optional argument for language (German = "DE" (default), English = "EN")}

\item{n.items}{optional argument to set the number of items per rank (default: all items; if fewer: items with highest values)}

\item{ncol.wrap}{optional argument to set the number of facet_wrap columns when distinguishing by subgroup or treatment group}

\item{textsize}{optional argument to adjust the text's size}

\item{min_textsize}{optional argument to set the minimum text size}

\item{percent_position}{optional argument to set position of percentage within bar ("center", "left", "top", etc.)}

\item{barwidth}{optional argument to define the width of the bar}

\item{legend.pos}{optional argument to set position of legend}
}
\value{
a horizontal barplot for multiple items
}
\description{
Vertical barplot to rank items
}
