% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sep_palette.R
\name{sep_palette}
\alias{sep_palette}
\title{ETH color palettes}
\usage{
sep_palette(name, n)
}
\arguments{
\item{name}{Name of the palette}

\item{n}{Number of numbers to be chosen from the palette}
}
\value{
ETH color palette of your choice
}
\description{
This is a non-exhaustive selection of official color palettes as listed on the ETH Zurich website. For further information, see: \href{https://ethz.ch/services/en/service/communication/corporate-design/colours.html}{ETH Colors}. This function was largely adopted from Karthik Ram's great \href{https://github.com/karthik/wesanderson}{wesanderson} package.
}
\examples{
sep_palette("Darkblues")
sep_palette("Pinks", n = 3)
}
