% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/septheme.R
\name{septheme}
\alias{septheme}
\title{The \code{SEP} theme}
\usage{
septheme(
  base_family = "IBM Plex Sans",
  base_size = 11.5,
  plot_title_family = base_family,
  plot_title_size = 18,
  plot_title_face = "bold",
  plot_title_margin = 10,
  subtitle_family = base_family,
  subtitle_size = 12,
  subtitle_face = "plain",
  subtitle_margin = 15,
  strip_text_family = base_family,
  strip_text_size = 12,
  strip_text_face = "plain",
  caption_family = base_family,
  axis_text_size = base_size,
  axis_title_family = subtitle_family,
  axis_title_size = 9,
  axis_title_face = "plain"
)
}
\arguments{
\item{base_family}{base font family}

\item{base_size}{base font size}

\item{plot_title_family}{plot title family}

\item{plot_title_size}{plot title size}

\item{plot_title_face}{plot title face}

\item{plot_title_margin}{plot title margin}

\item{subtitle_family}{subtitle font family}

\item{subtitle_size}{subtitle size}

\item{subtitle_face}{subtitle face}

\item{subtitle_margin}{subtitle margin}

\item{strip_text_family}{strip text family}

\item{strip_text_size}{strip text size}

\item{strip_text_face}{strip text face}

\item{caption_family}{caption font family}

\item{axis_text_size}{axis text size}

\item{axis_title_family}{axis title font family}

\item{axis_title_size}{axis title size}

\item{axis_title_face}{axis title face}
}
\value{
the sep theme
}
\description{
The \code{SEP} theme
}
\examples{
\dontrun{

## set as theme
theme_set(septheme())

## or use locally

library(ggplot2)
ggplot(data = mtcars, aes(x = gear, y = am)) +
geom_point() +
septheme()
}
}
