% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share_dk.R
\name{share_dk}
\alias{share_dk}
\title{Calculate share of "Don't know"-answers for survey item}
\usage{
share_dk(data, item)
}
\arguments{
\item{data}{a \code{data.frame} object}

\item{item}{relevant survey item (i.e. w5_q17)}
}
\value{
relative frequency of "Don't know"-answers, if applicable to item
}
\description{
Calculate share of "Don't know"-answers for survey item
}
