% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_barplot.R
\name{simple_barplot}
\alias{simple_barplot}
\title{Simple barplot}
\usage{
simple_barplot(df, var, title, ..., barcolor = "#1F407A")
}
\arguments{
\item{df}{dataframe}

\item{var}{variable}

\item{title}{plot title}

\item{...}{any other option for geom_bar}

\item{barcolor}{color of the bar}
}
\value{
a plot
}
\description{
Creates a simple barplot for any variable.
}
\examples{
simple_barplot(df = mtcars, var = gear, title = "Gears", width = 0.3)
}
