% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_sumtab.R
\name{cb_sumtab}
\alias{cb_sumtab}
\title{Summary Table of survey data for Codebook}
\usage{
cb_sumtab(metadata, response, num.var, na_sep = TRUE, type)
}
\arguments{
\item{type}{}
}
\value{
summary table for factor or numeric variables
}
\description{
Summary Table of survey data for Codebook
}
