% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_tab.R
\name{cb_tab}
\alias{cb_tab}
\title{Creates a Codebook Table for a given variable (row)}
\usage{
cb_tab(
  data = upanel.meta,
  num.var,
  .meta = meta,
  .codes = codes,
  .codes.en = codes.en,
  .miscodes = miscodes,
  .str_dk = str_dk,
  .str_no = str_no,
  lbl.space = "1em",
  lblen.space = "1em",
  mis.space = "1em"
)
}
\arguments{
\item{data}{data placeholder}

\item{num.var}{num.var placeholder}

\item{.meta}{meta placeholder}

\item{.codes}{codes placeholder}

\item{.codes.en}{codes.en placeholder}

\item{.miscodes}{miscodes placeholder}

\item{.str_dk}{str_dk placeholder}

\item{.str_no}{str_no placeholder}

\item{lbl.space}{lbl.space placeholder}

\item{lblen.space}{lblen.space placeholder}

\item{mis.space}{mis.space placeholder}
}
\value{
Table from Kable in Latex Format
}
\description{
Creates a Codebook Table for a given variable (row)
}
