% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_table.R
\name{cb_tab}
\alias{cb_tab}
\title{Codebook Table for one variable}
\usage{
cb_tab(
  metadata,
  num.var,
  .meta = meta,
  .codes = codes,
  .codes.en = codes.en,
  .miscodes = miscodes,
  lbl.space = "1em",
  lblen.space = "1em",
  mis.space = "1em"
)
}
\arguments{
\item{metadata}{a \code{data.frame} object with survey metadata}

\item{num.var}{single numeric value indexing column of variable in metadata}

\item{.meta}{numeric vector indexing rows of variable information in metadata (defaults to meta)}

\item{.codes}{numeric vector indexing rows of German value labels in metadata (defaults to codes)}

\item{.codes.en}{numeric vector indexing rows of English value labels in metadata (defaults to codes.en)}

\item{.miscodes}{numeric vector indexing rows of missing labels in metadata (defaults to miscodes)}

\item{lbl.space}{optional argument to adjust table space between variable information and German value labels (defaults to 1em)}

\item{lblen.space}{optional argument to adjust table space between German value labels and English value labels (defaults to 1em)}

\item{mis.space}{optional argument to adjust table space between English value labels and missing labels (defaults to 1em)}
}
\value{
codebook table in Latex format
}
\description{
Codebook Table for one variable
}
