% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_tab_multi.R
\name{cb_tab_multi}
\alias{cb_tab_multi}
\title{Creates Page with Table in Codebook for multiple variabes}
\usage{
cb_tab_multi(
  multi.vars,
  data = upanel.meta,
  .cbtab = cbtab,
  comment = "",
  lbl.space = "1em",
  lblen.space = "1em",
  mis.space = "1em"
)
}
\arguments{
\item{multi.vars}{multi.vars placeholder}

\item{data}{a \code{data.frame} object}

\item{.cbtab}{.cbtab placeholder}

\item{comment}{comment placeholder}

\item{lbl.space}{lbl.space placeholder}

\item{lblen.space}{lblen.space placeholder}

\item{mis.space}{mis.space placeholder}

\item{num.var}{num.var placeholder}
}
\value{
cb_tab_multi object
}
\description{
Creates Page with Table in Codebook for multiple variabes
}
