% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_par.R
\name{n_par}
\alias{n_par}
\title{Number of participants for survey item}
\usage{
n_par(data, item, lang = "DE")
}
\arguments{
\item{data}{a \code{data.frame} object}

\item{item}{relevant survey item (i.e. w5_q17)}

\item{lang}{language (German = "DE" (default), English = "EN", anything else goes neutral)}
}
\value{
Number of participants ("Don't knows"-answers included) for item
}
\description{
Number of participants for survey item

To be used with \code{plot_bar_h()}, \code{plot_bar_v()}, \code{plot_groupbar_h()} and \code{plot_groupbar_v()}.

For \code{plot_multiple_h()} and  \code{plot_rank_v()}, use the function \code{n_par_by()}.
}
