% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_par.R
\name{n_par}
\alias{n_par}
\title{Number of participants for survey item}
\usage{
n_par(data, item)
}
\arguments{
\item{data}{a \code{data.frame} object}

\item{item}{relevant survey item (i.e. w5_q17)}
}
\value{
number of participants ("Don't knows"-answers included) for item
}
\description{
Number of participants for survey item
}
