% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_par_by.R
\name{n_par_by}
\alias{n_par_by}
\title{Number of participants for survey item}
\usage{
n_par_by(data, item, by = NULL, treat = NULL, lang = "DE")
}
\arguments{
\item{data}{a \code{data.frame} object}

\item{item}{relevant survey item (i.e. w5_q17)}

\item{by}{optional argument for distinction by subgroup}

\item{treat}{optional argument for distinction by treatment group}

\item{lang}{language (German = "DE" (default), English = "EN", anything else goes neutral)}
}
\value{
Number of participants ("Don't knows"-answers included) for item
}
\description{
Number of participants for survey item, distinguishable by subgroup and/or treatment group.

To be used with \code{plot_multiple_h()} and  \code{plot_rank_v()}.

For \code{plot_bar_h()}, \code{plot_bar_v()}, \code{plot_groupbar_h()} and \code{plot_groupbar_v()}, use the function \code{n_par()}.
}
