% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar_h.R
\name{plot_bar_h}
\alias{plot_bar_h}
\title{Simple horizontal barplot}
\usage{
plot_bar_h(
  data = currentwave,
  item,
  dk_share = "?",
  barcolor = "#1F407A",
  barwidth = 0.8,
  texthjust = 1.2,
  textsize = 4,
  textcolor = "white"
)
}
\arguments{
\item{data}{a \code{data.frame} object}

\item{item}{a survey item}

\item{dk_share}{share of "Don't knows". Check \code{eda_freqtable} for share.}

\item{barcolor}{optional argument to set the color of the bar}

\item{barwidth}{optional argument to define the width of the bar}

\item{texthjust}{optional argument to adjust the text's horizontal alignment}

\item{textsize}{optional argument to adjust the text's size}

\item{textcolor}{optional argument to adjust the text's colour (default to white)}
}
\value{
a horizontal barplot
}
\description{
Simple horizontal barplot
}
