% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar_v.R
\name{plot_bar_v}
\alias{plot_bar_v}
\title{Simple vertical barplot}
\usage{
plot_bar_v(
  data,
  item,
  barcolor = "#1F407A",
  barwidth = 0.8,
  textvjust = 2,
  textsize = 3.5
)
}
\arguments{
\item{data}{a \code{data.frame} object}

\item{item}{a survey item}

\item{barcolor}{optional argument to set the color of the bar}

\item{barwidth}{optional argument to define the width of the bar}

\item{textvjust}{optional argument to adjust the text's horizontal alignment}

\item{textsize}{optional argument to adjust the text's size}
}
\value{
a vertical barplot
}
\description{
Simple vertical barplot
}
