% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_groupbar_h.R
\name{plot_groupbar_h}
\alias{plot_groupbar_h}
\title{Grouped horizontal barplot}
\usage{
plot_groupbar_h(
  data,
  item,
  by,
  lang = "DE",
  barpadding = 0.1,
  barposition = "dodge",
  legendtitle = "",
  textsize = 8,
  min_textsize = 5,
  ...
)
}
\arguments{
\item{data}{a \code{data.frame} object}

\item{item}{a survey item}

\item{by}{grouping variable}

\item{lang}{optional argument for language (German = "DE" (default), English = "EN")}

\item{barpadding}{optional argument to adjust padding between bars}

\item{barposition}{optional argument to determine the positioning of the bars (default: "dodge")}

\item{legendtitle}{optional argument to define a legend title}

\item{textsize}{optional argument to adjust the text's size}

\item{min_textsize}{optional argument to set the minimum text size}

\item{...}{further arguments of \code{scale_fill_manual} like \code{labels}}
}
\value{
a grouped horizontal barplot
}
\description{
Grouped horizontal barplot
}
