% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_groupbar_v.R
\name{plot_groupbar_v}
\alias{plot_groupbar_v}
\title{Grouped vertical barplot}
\usage{
plot_groupbar_v(
  data,
  item,
  by,
  barpadding = 0.1,
  barwidth = 0.5,
  legendtitle = "",
  textsize = 8,
  min_textsize = 5,
  ...
)
}
\arguments{
\item{data}{a \code{data.frame} object}

\item{item}{a survey item}

\item{by}{grouping variable}

\item{barpadding}{optional argument to adjust padding between bars}

\item{barwidth}{optional argument to define the bar's width}

\item{legendtitle}{optional argument to define a legend title}

\item{textsize}{optional argument to adjust the text's size}

\item{min_textsize}{optional argument to set the minimum text size}

\item{...}{further arguments of \code{scale_fill_manual} like \code{labels}}
}
\value{
a grouped vertical barplot
}
\description{
Grouped vertical barplot
}
