% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_barplot.R
\name{simple_barplot}
\alias{simple_barplot}
\title{Create simple barplots}
\usage{
simple_barplot(df, var, title, ...)
}
\arguments{
\item{df}{dataframe}

\item{var}{variable}

\item{title}{plot title}

\item{...}{any other option for geom_bar}
}
\value{

}
\description{
Create simple barplots
}
\examples{
simple_barplot(df = mtcars, var = gear, title = "Gears", width = 0.3)
}
